refactor: move svelte client files to svelte folder root
