+++ examples/svelte/.gitignore
