+++ examples/svelte/.npmrc
