+++ examples/svelte/package.json
