+++ examples/svelte/src/app.d.ts
