+++ examples/svelte/static/favicon.png
