+++ examples/svelte/svelte.config.js
