+++ examples/svelte/vite.config.ts
